Start moving test-related functionality
